// hr-frontend/src/lib/api/generated/model/employeeStatusSetting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface EmployeeStatusSetting {
  employeeStatusId?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  description?: string;
  active?: boolean;
  usageCount?: number;
  createdAt?: string;
  updatedAt?: string;
}